+++ test/Sentry.Tests/Integrations/AppDomainAdapterTests.cs
@@ -9,26 +9,22 @@ namespace Sentry.Tests.Integrations
 {
     public class AppDomainAdapterTests
     {
-        [Fact]
+        [SkippableFact]
         public void UnhandledException_FiredOnExceptionUnhandledInThread()
         {
             // Test flaky on Mono
-            if (!Runtime.Current.IsMono())
+            Skip.If(Runtime.Current.IsMono());
+
+            var evt = new ManualResetEventSlim(false);
+            AppDomainAdapter.Instance.UnhandledException += (_, _) => evt.Set();
+
+            var thread = new Thread(() => throw new Exception())
             {
-                var evt = new ManualResetEventSlim(false);
-                AppDomainAdapter.Instance.UnhandledException += (_, _) =>
-                {
-                    evt.Set();
-                };
+                IsBackground = false
+            };
 
-                var thread = new Thread(() => throw new Exception())
-                {
-                    IsBackground = false
-                };
-
-                thread.Start();
-                Assert.True(evt.Wait(TimeSpan.FromSeconds(3)));
-            }
+            thread.Start();
+            Assert.True(evt.Wait(TimeSpan.FromSeconds(3)));
         }
     }
 }