+++ test/Sentry.Tests/Protocol/Context/DeviceTests.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Sentry.Tests.Helpers;
 using Xunit;
 
 // ReSharper disable once CheckNamespace
@@ -20,6 +21,13 @@ namespace Sentry.Protocol.Tests.Context
         [Fact]
         public void SerializeObject_AllPropertiesSetToNonDefault_SerializesValidObject()
         {
+            var timeZone = TimeZoneInfo.CreateCustomTimeZone(
+                "tz_id",
+                TimeSpan.FromHours(2),
+                "my timezone",
+                "my timezone"
+            );
+
             var sut = new Device
             {
                 Name = "testing.sentry.io",
@@ -43,7 +51,7 @@ namespace Sentry.Protocol.Tests.Context
                 ModelId = "0921309128012",
                 Orientation = DeviceOrientation.Portrait,
                 Simulator = false,
-                Timezone = TimeZoneInfo.Local,
+                Timezone = timeZone,
                 UsableMemory = 100,
                 LowMemory = true
             };
@@ -52,8 +60,8 @@ namespace Sentry.Protocol.Tests.Context
 
             Assert.Equal(
                 "{\"type\":\"device\"," +
-                $"\"timezone\":\"{TimeZoneInfo.Local.Id}\"," +
-                $"\"timezone_display_name\":\"{TimeZoneInfo.Local.DisplayName.Replace("+", "\\u002B")}\"," +
+                "\"timezone\":\"tz_id\"," +
+                "\"timezone_display_name\":\"my timezone\"," +
                 "\"name\":\"testing.sentry.io\"," +
                 "\"manufacturer\":\"Manufacturer\"," +
                 "\"brand\":\"Brand\"," +